Mouse: Move interfaces to interfaces/ subfolder
