+++ xbmc/input/mouse/CMakeLists.txt
@@ -1,9 +1,9 @@
 set(SOURCES MouseWindowingButtonMap.cpp)
 
-set(HEADERS IMouseButtonMap.h
-            IMouseDriverHandler.h
-            IMouseInputHandler.h
-            IMouseInputProvider.h
+set(HEADERS interfaces/IMouseButtonMap.h
+            interfaces/IMouseDriverHandler.h
+            interfaces/IMouseInputHandler.h
+            interfaces/IMouseInputProvider.h
             MouseWindowingButtonMap.h)
 
 core_add_library(input_mouse)
