+++ xbmc/input/mouse/interfaces/IMouseButtonMap.h
