+++ xbmc/input/mouse/interfaces/IMouseDriverHandler.h
