+++ xbmc/input/mouse/interfaces/IMouseInputHandler.h
