+++ xbmc/input/mouse/interfaces/IMouseInputProvider.h
